requirements.txt
# Core Framework
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1

# PDF Parsing & Document Processing
docling==2.62.0
python-magic==0.4.27
pypdf==5.1.0


# LLM & AI Frameworks
langchain==0.3.16
langchain-community==0.3.16
langchain-openai==0.2.14
langgraph==0.2.60
openai==1.59.7

# Vector Database & Search
pymilvus==2.5.2
elasticsearch==8.17.0
sentence-transformers==3.3.1

# Database Clients (using psycopg3 instead of psycopg2-binary)
psycopg[binary]>=3.2.10
pgvector==0.3.6
snowflake-connector-python==3.12.3


# API Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1

# Observability & Monitoring
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
prometheus-client==0.21.1

# Utilities
aiofiles==24.1.0
tenacity==9.0.0
tiktoken==0.8.0
numpy==2.2.1
pandas==2.2.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
locust==2.32.4
faker==33.1.0

# Development
black==24.10.0
ruff==0.8.5
mypy==1.14.0
pre-commit==4.0.1

